update readme, move tests, re enable test runner
