+++ README.md
@@ -40,6 +40,11 @@ You will also see any lint errors in the console.
 ### `npm test`
 
 Launches the test runner in the interactive watch mode.<br>
+
+The unit test scripts are under `./__tests__` folder and tested functions are under `./src/store`.
+
+The tested functions are all the actions of feed, message, plugin, ui, and user.
+
 See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
 
 ### `npm run build`
@@ -61,10 +66,6 @@ If you aren’t satisfied with the build tool and configuration choices, you can
 Instead, it will copy all the configuration files and the transitive dependencies (Webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
 
 You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
-
-The unit test scripts are under `./__tests__` folder and tested functions are under `./src/store`.
-
-The tested functions are all the actions of feed, message, plugin, ui, and user.
 
 ## Learn More
