+++ package.json
@@ -67,21 +67,5 @@
     "keymirror": "^0.1.1",
     "redux-logger": "^3.0.6",
     "tslint": "^5.13.1"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "jsx",
-      "json",
-      "ts",
-      "tsx"
-    ],
-    "testMatch": [
-      "**/__tests__/**/*.(js|ts)?(x)",
-      "**/?(*.)(spec|test).(js|ts)?(x)"
-    ],
-    "transform": {
-      "^.+\\.tsx?$": "babel-jest"
-    }
   }
 }
